[TASK] Add missing "description" field for composer.json
The "description" field in `composer.json` is mandantory. This
information does not exits in this package `composer.json`
file.

Newer composer versions seems to handle this pretty well, e.g.
if the `show --tree` command is executed. But quite a bunch
version will fail with a `TypeError` exception.

Therefore, a simple "description" is added now to mitigate
this issue.

Resolves: #9